Improve clone/checkout/build logic
Now lettuce checks if a clone already exists and if it does, if its HEAD
matches the head of the requested repo and branch.  This means if you
change either the branch or the repo url in the config file, lettuce
will update the checkout and rebuild the model.  Similarly, even if you
haven't changed those things, but the version on the remote repo *has*
changed, lettuce will update the checkout and rebuild the model.

lettuce will always checkout the desired branch as a detached head and
leave a single remote called origin that is the source of that detached
head.

If the repo directory does not exist or does not have a valid git repo
in it, a fresh clone is made.

This logic should give the desired behavior for both automated nighly
testing mode and 'interactive' manual sessions with lettuce.

I've tested most of the possible combinations and it appears to work as
advertised.  It does write a bit more information to stdout than before,
but personally I like to know what it did to the repo, if anything.